backend/requirements.txt
django==3.0.3
tensorflow==2.0.0
tensorflow_hub==0.7.0
mtcnn==0.1.0
djangorestframework==3.11.0
Pillow==7.0.0
numpy==1.18.1
scipy==1.4.1
tweepy==3.8.0
stop-words==2018.7.23
eli5==0.10.1

